requirement.txt
# ============================================================
# PacTravel Data Warehouse - ELT Pipeline
# Requirements File
# ============================================================

# --------------------------
# Extract & Load (Luigi)
# --------------------------
luigi
pandas
SQLAlchemy
psycopg2-binary

# --------------------------
# Scheduling
# --------------------------
schedule

# --------------------------
# Alerting
# --------------------------
sentry-sdk
# --------------------------
# Utilities
# --------------------------
python-dotenv
loguru